api/requirements.txt
argcomplete==3.2.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.34.15
botocore==1.34.15
certifi==2023.11.17
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.9
django-debug-toolbar==4.2.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
durationpy==0.6
hjson==3.1.0
idna==3.6
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
mysqlclient==2.2.1
oauthlib==3.2.2
placebo==0.9.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.3
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
zappa==0.58.0
